Refactor RollingEnv as described in commit message
1. Change step to return values in accordance with base class
2. Allow for default resetting the environment
3. Return heightmap as obs by default (can change later)
4. Moved heightmap generation functions to rolling env file
5. Reorganize tests and add pytest w/ test script
