+++ examples/dough_model_learning/test/test_data_generation.py
@@ -8,39 +8,6 @@ import torch
 import mani_skill2.envs.mpm.rolling_env as rolling_env
 
 
-class TestGenerateDomeHeightmap(unittest.TestCase):
-    def test(self):
-        dome_radius = 0.1
-        dome_height = 0.05
-        dx = 0.001
-        height_map = data_generation.generate_dome_heightmap(
-            dome_radius,
-            dome_height,
-            dx,
-        )
-        self.assertAlmostEqual(height_map.max(), dome_height)
-        self.assertEqual(height_map.min(), 0)
-
-        # Test the height at a random position.
-        x_index = 70
-        y_index = 90
-        half_width = int(dome_radius / dx)
-        xy_coordinate = np.array(
-            [(half_width - x_index) * dx, (half_width - y_index) * dx]
-        )
-        xy_height = height_map[x_index, y_index]
-
-        sphere_radius = (dome_radius ** 2 + dome_height ** 2) / (2 * dome_height)
-        self.assertAlmostEqual(
-            (sphere_radius - dome_height) ** 2 + dome_radius ** 2, sphere_radius ** 2
-        )
-        self.assertAlmostEqual(
-            (xy_height + (sphere_radius - dome_height)) ** 2
-            + (xy_coordinate ** 2).sum(),
-            sphere_radius ** 2,
-        )
-
-
 class TestGenerateTransitionTuple(unittest.TestCase):
     def test(self):
         env = rolling_env.RollingEnv(sim_freq=500, mpm_freq=2000)
